+++ test/unit/test_runner.test.ts
@@ -1,9 +1,9 @@
 import { expect, test } from 'bun:test'
 import { unwrap } from '@serkonda7/ts-result'
 import { ScopeRegexMode, TestRunner } from '../../src/unit/index.ts'
-import { read_data, TESTLANG_GRAMMARS } from '../testutil.ts'
+import { REGISTRY, read_data } from '../testutil.ts'
 
-const runner = new TestRunner(TESTLANG_GRAMMARS, ScopeRegexMode.standard)
+const runner = new TestRunner(REGISTRY, ScopeRegexMode.standard)
 
 test('OK on correct grammar test', async () => {
 	const res = unwrap(await runner.test_file(read_data('parser.testlang')))